slack_bot_v2/pyproject.toml
[tool.poetry]
name = "slack-bot-v2"
version = "0.1.0"
description = ""
authors = ["Arc1el <khm970514@gmail.com>"]
readme = "README.md"
packages = [{include = "slack_bot_v2"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
slackclient = "^2.9.4"
python-dotenv = "^0.21.1"
uvicorn = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "slack_bot_v2.main:start"